Make internal things internal
