+++ Sources/ForageSDK/Component/ForagePANTextField/MaskedUITextField.swift
@@ -9,13 +9,7 @@
 import Foundation
 import UIKit
 
-extension String {
-    subscript(safe index: Int) -> Character? {
-        index < count && index >= 0 ? self[self.index(startIndex, offsetBy: index)] : nil
-    }
-}
-
-enum MaskPattern: String {
+internal enum MaskPattern : String {
     case unset = "###################"
     case sixteenDigits = "#### #### #### ####"
     case eighteenDigits = "###### #### ##### ## #"
