+++ Tests/ForageSDKTests/ForagePANTextFieldTests.swift
@@ -81,15 +81,7 @@ final class ForagePANTextFieldTests: XCTestCase {
 
         XCTAssertFalse(changesAllowed)
     }
-
-    func test_elementHeight() {
-        let newHeight = 100.0
-        foragePANTextField.elementHeight = newHeight
-
-        let height = foragePANTextField.elementHeight
-        XCTAssertEqual(newHeight, height)
-    }
-
+    
     func test_font() {
         let newFont = UIFont.systemFont(ofSize: 14, weight: .regular)
         foragePANTextField.font = newFont
